Remove old functions and code to make sure each class has one clear responsiblity
