+++ Engines/engine.py
@@ -1,11 +1,13 @@
 from random import choice
-from game import Game
+from GameClasses.game import Game
 from Moves.moveFactory import MoveFactory
+from Moves.moveGenerator import MoveGenerator
 
 class Engine():
     def __init__(self):
         self.game = Game()
         self.move_factory = MoveFactory()
+        self.move_generator = MoveGenerator()
 
     def set_position(self, fenstr, moves):
         self.game.setup_fenstr(fenstr)
@@ -16,7 +18,7 @@ class Engine():
             self.game.make_move(move)
 
     def go(self):
-        moves = self.game.get_valid_moves()
+        moves = self.move_generator.get_all_moves()
 
         if len(moves) > 0:
             return choice(moves)
@@ -26,15 +28,9 @@ class Engine():
     def print_game_state(self):
         self.game.display_game()
 
+        moves = self.move_generator.get_all_moves()
 
-        eval = self.game.evaluate_state(0)
-        print(f"{eval=}")
-
-        moves = self.game.get_valid_moves()
         print(f"{moves=}")
-        is_checkmate = self.game.is_checkmate()
-        print(f"{is_checkmate}")
-
