+++ GameClasses/board.py
@@ -5,15 +5,13 @@ from strings import *
 from enums import *
 from coords import Coords
 
-from Moves.move import Move
-
 from Pieces.pieceFactory import PieceFactory
+
 
 class Board:
     def __init__(self):
         self._piece_factory = PieceFactory()
         self._board = [[None]*8 for _ in range(8)]
-        self._previous_moves = []
 
     def get_square(self, coords: Coords) -> Piece:
         return self._board[8-coords.rank][coords.file.value]
@@ -47,73 +45,62 @@ class Board:
         print()
         print(f"  {['{:^3}'.format(File(i).name) for i in range(8)]}")
 
-    def get_moves(self, player: Colour, enpassant_coords: Coords):
-        moves = []
+    # def get_moves(self, gamestate):
+    #     moves = []
 
-        for coords in self._all_squares_iterator():
-            piece = self.get_square(coords)
+    #     for coords in self._all_squares_iterator():
+    #         piece = self.get_square(coords)
 
-            if piece and piece.colour == player:
-                moves.extend(self.get_piece_moves(piece, coords))
+    #         if piece and piece.colour == player:
+    #             moves.extend(self.get_piece_moves(piece, coords))
 
-        return moves 
+    #     return moves 
 
-    def get_piece_moves(self, piece, coords):
-        moves = []
+    # def get_piece_moves(self, piece, coords):
+    #     moves = []
 
-        candidate_moves = piece.get_candidate_moves(coords)
+    #     candidate_moves = piece.get_candidate_moves(coords)
 
-        for candidate in candidate_moves:
-            for end_coords in candidate.generate_coords(coords):
-                if self.get_square(end_coords) == None:
-                    if candidate.capture_forced:
-                        break
-                    moves.append(Move(piece.colour, coords, False, end_coords, False))
+    #     for candidate in candidate_moves:
+    #         for end_coords in candidate.generate_coords(coords):
+    #             if self.get_square(end_coords) == None:
+    #                 if candidate.capture_forced:
+    #                     break
+    #                 moves.append(Move(piece.colour, coords, False, end_coords, False))
 
-                else:
-                    blocking_piece = self.get_square(end_coords)
-                    if candidate.capture_allowed and blocking_piece.colour != piece.colour:
-                        moves.append(Move(piece.colour, coords, True, end_coords, False))
-                    break
+    #             else:
+    #                 blocking_piece = self.get_square(end_coords)
+    #                 if candidate.capture_allowed and blocking_piece.colour != piece.colour:
+    #                     moves.append(Move(piece.colour, coords, True, end_coords, False))
+    #                 break
 
-        return moves
+    #     return moves
 
     # def make_castle_move(self, move):
     #     if move.end_coords == Coords.init_from_str("f1"):
             
 
 
+    # def make_move(self, move: Move):
+    #     piece = self.get_square(move.start_coords)
+    #     end_piece = self.get_square(move.end_coords)
 
-    # Assumes the move is valid for now
-    def make_move(self, move: Move):
-        piece = self.get_square(move.start_coords)
-        end_piece = self.get_square(move.end_coords)
+    #     move.end_piece = end_piece
+    #     move.has_piece_moved_before = piece.has_moved
 
-        move.end_piece = end_piece
-        move.has_piece_moved_before = piece.has_moved
+    #     self._previous_moves.append(move)
 
-        self._previous_moves.append(move)
+    #     piece.has_moved = True
+    #     self.set_square(None, move.start_coords)
+    #     self.set_square(piece, move.end_coords)
 
-        piece.has_moved = True
-        self.set_square(None, move.start_coords)
-        self.set_square(piece, move.end_coords)
-
-        if type(piece) == King and \
-            abs(move.start_coords.file.value - move.end_coords.file.value) == 2:
-            self.make_castle_move(move)
+    #     if type(piece) == King and \
+    #         abs(move.start_coords.file.value - move.end_coords.file.value) == 2:
+    #         self.make_castle_move(move)
 
 
-    def undo_last_move(self):
-        if len(self._previous_moves) == 0:
-            print("No moves. found")
-            return
-        last_move = self._previous_moves.pop()
-
-        piece = self.get_square(last_move.end_coords)
-        piece.has_moved = last_move.has_piece_moved_before
-
-        self.set_square(piece, last_move.start_coords)
-        self.set_square(last_move.end_piece, last_move.end_coords)
+    def undo_last_move(self, move):
+        return
 
     def is_player_in_check(self, player: Colour):
         player_in_check = False
@@ -142,16 +129,16 @@ class Board:
 
         return player_in_check
 
-    def is_player_left_in_check(self, move: Move):
-        # Temporarily make move
-        self.make_move(move)
+    # def is_player_left_in_check(self, move: Move):
+    #     # Temporarily make move
+    #     self.make_move(move)
 
-        player_in_check = self.is_player_in_check(move.player_to_move)
+    #     player_in_check = self.is_player_in_check(move.player_to_move)
         
-        # Undo the move
-        self.undo_last_move()
+    #     # Undo the move
+    #     self.undo_last_move()
 
-        return player_in_check
+    #     return player_in_check
             
         
     # def _get_all_player_pieces(self, player: Colour):
