+++ Moves/moveGenerator.py
@@ -1,0 +1,5 @@
+from GameClasses.game import Game
+
+class MoveGenerator():
+    def get_all_moves(game: Game):
+        return []
