+++ main.py
@@ -1,7 +1,7 @@
 from uci import Uci
 
 from Engines.engine import Engine
-from game import Game
+from GameClasses.game import Game
 
 from Moves.moveFactory import MoveFactory
